init commit with some modification
